created separate subprojects
