+++ cli/build.gradle
@@ -1,0 +1,6 @@
+version = "1.0"
+
+dependencies {
+  compile project(":sum")
+}
+
