+++ settings.gradle
@@ -1,2 +1,4 @@
 rootProject.name = 'intro17'
 
+include 'cli', 'sum'
+
