+++ sum/build.gradle
@@ -1,0 +1,2 @@
+version = '1.0'
+
