+++ sum/src/main/java/Sum.java
@@ -1,0 +1,10 @@
+public class Sum {
+  static int sum(int[] arr) {
+    int ans = 0;
+    for (int i = 0; i < arr.length; i++) {
+      ans += arr[i];
+    }
+    return ans;
+  }
+}
+
